// backend/src/types/payload-types.ts
export interface Destination {
    id: string;
    type: 'Feature';
    /**
     * @minItems 2
     * @maxItems 2
     */
    geometry?: [number, number] | null;
    properties: {
        mapbox_id?: string | null;
        osm_id?: string | null;
        osm_type?: string | null;
        wikidata_id?: string | null;
        geonames_id?: number | null;
        wikipedia_url?: string | null;
        destination_type: 'poi' | 'address' | 'area';
        name: string;
        description?: string | null;
        address?: string | null;
        main_photo?: (string | null) | Media;
        poi_category?: string | null;
        timezone?: string | null;
        bbox?: {
            min_latitude?: number | null;
            min_longitude?: number | null;
            max_latitude?: number | null;
            max_longitude?: number | null;
        };
    };
    updatedAt: string;
    createdAt: string;
}

/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
    id: string;
    image_origin: 'trek' | 'external';
    artist?: string | null;
    license_url?: string | null;
    license?: string | null;
    updatedAt: string;
    createdAt: string;
    url?: string | null;
    filename?: string | null;
    mimeType?: string | null;
    filesize?: number | null;
    width?: number | null;
    height?: number | null;
    focalX?: number | null;
    focalY?: number | null;
    sizes?: {
        thumbnail?: {
            url?: string | null;
            width?: number | null;
            height?: number | null;
            mimeType?: string | null;
            filesize?: number | null;
            filename?: string | null;
        };
        card?: {
            url?: string | null;
            width?: number | null;
            height?: number | null;
            mimeType?: string | null;
            filesize?: number | null;
            filename?: string | null;
        };
        tablet?: {
            url?: string | null;
            width?: number | null;
            height?: number | null;
            mimeType?: string | null;
            filesize?: number | null;
            filename?: string | null;
        };
    };
}